backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pinecone-client==3.2.2
google-genai>=0.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.3.1
pypdf==4.2.0
aiosqlite==0.20.0
databases==0.9.0
alembic==1.13.1
ruff==0.4.7
pytest==8.2.2
pytest-asyncio==0.23.7
